# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server vote server.",
        "title": "vote",
        "contact": {
            "email": "2695062156@qq.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "执行用户登录操作",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "执行用户登录",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码ID",
                        "name": "captcha_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码值",
                        "name": "captcha_value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "description": "执行用户退出",
                "tags": [
                    "login"
                ],
                "summary": "执行用户退出",
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            },
            "post": {
                "description": "创建一个新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "创建一个新用户",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "确认密码",
                        "name": "password2",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        },
...
        "/vote": {
            "get": {
                "description": "获取特定投票信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vote"
                ],
                "summary": "获取特定投票信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tools.ECode"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "tools.ECode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}